aqi-karachi/cicd/requirements_cicd.txt
# Core dependencies for CI/CD
pymongo>=4.6.1
python-dotenv>=1.0.0
schedule>=1.2.0

# Your project dependencies (matching your requirements.txt)
streamlit==1.30.0
pandas==2.1.4
numpy==1.26.0
plotly==5.18.0
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2
statsmodels==0.14.1
seaborn==0.13.0
matplotlib==3.8.2
requests==2.31.0
openmeteo-requests==1.1.0
requests-cache==1.1.0
retry-requests==2.0.0

# Additional CI/CD utilities
python-dateutil>=2.8.2
pytz>=2023.3